requirements.txt
cffi==2.0.0
cryptography==46.0.2
pycparser==2.23
pycryptodome==3.23.0
typing_extensions==4.15.0
